requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
streamlit==1.35.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0

# AI / LLM
openai==1.30.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
langchain-core==0.2.3

# Vector DB
faiss-cpu==1.8.0
sentence-transformers==3.0.0

# Data Processing
pandas==2.2.2
numpy==1.26.4
beautifulsoup4==4.12.3
requests==2.32.3
httpx==0.27.0
lxml==5.2.2

# Document Generation
reportlab==4.2.0
fpdf2==2.7.9

# Utilities
tenacity==8.3.0
rich==13.7.1
loguru==0.7.2
python-multipart==0.0.9
aiohttp==3.9.5
tiktoken==0.7.0
